+++ CHANGELOG.md
@@ -7,11 +7,18 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
-[Unreleased]: https://github.com/atomist/rug/compare/0.7.1...HEAD
+[Unreleased]: https://github.com/atomist/rug/compare/0.8.0...HEAD
+
+## [0.8.0] - 2017-01-04
+
+[0.8.0]: https://github.com/atomist/rug/compare/0.7.1...0.8.0
 
 ### Added
 
--   NavigationAxis: New AxisSpecifier that allows navigating via a property in path expression language
+-   NavigationAxis: New AxisSpecifier that allows navigating via a
+    property in path expression language
+
+-   Path expression predicates can contain path expressions
 
 ### Changed
 
@@ -25,6 +32,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 -   **BREAKING** Rug DSL: A Rug can no longer specify its action in JavaScript:
     In this case, use TypeScript or JavaScript.
 
+-   **BREAKING** FileArtifactMutableView is now named FileMutableView
 
 ### Fixed
 
@@ -34,7 +42,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 -   Double-quoted strings in Rug DSL are now interpreted similarly to
     Java double-quoted strings
 
--   LABEL section in files parsed by Dockerfile type did not handle 
+-   LABEL section in files parsed by Dockerfile type did not handle
     multi-line strings correctly as per https://github.com/atomist/rug/issues/140
 
 ## [0.7.1] - 2016-12-19